now it works

+++ ft_strncat.c
@@ -14,8 +14,8 @@
 
 char	*ft_strncat(char *s1, const char *s2, size_t n)
 {
-	size_t	len;
-	int		i;
+	size_t len;
+	size_t i;
 
 	len = ft_strlen(s1);
 	i = 0;
